backend/requirements.txt
flask
google-genai
python-dotenv
gunicorn
flask-login
flask-bcrypt
flask-sqlalchemy
